Add sorting function to internal list
Currently, the app does not sort the names on the display when
adding new sellers/buyers or even when editing their names. The new
contacts added are just appended at the end.

By sorting the list when add or edit command is executed, we can
ensure that there's no inconsistency in the order of the contacts
listed.