Fix documentation bug
Currently, there are no constraints on our index value and budgets.
This could be misleading to users as exceeding these values may
cause unexpected behaviour in the application.

By adding these constraints to the documentation, we can warn the
users to be wary of adding values that exceed the constraints.